Introduce `shared` module
