+++ protocols/stream/src/lib.rs
@@ -3,6 +3,7 @@
 mod behaviour;
 mod control;
 mod handler;
+mod shared;
 mod upgrade;
 
 pub use behaviour::{AlreadyRegistered, Behaviour};
